idiots_geniuses/requirements.txt
python-dotenv==0.7.1
tweepy==3.5.0